src/business/pattern.ts
// src/business/pattern.ts
import { RuleManager } from './rule';
import {
    PatternDefinition,
    PatternRules,
    Position,
    ValidationResult
} from './types/pattern.types';

export class MusicSetPattern {
  private ruleManager: RuleManager;
  private definition: PatternDefinition;
  private patternRules: PatternRules;

  constructor(definition: PatternDefinition) {
    this.ruleManager = RuleManager.getInstance();
    this.definition = definition;
    this.patternRules = this.ruleManager.getPatternRules();
  }

  public isValid(): ValidationResult {
    // Implementation will go here
  }

  public isValidForPosition(position: Position): ValidationResult {
    // Implementation will go here
  }

  public validateCategoryPlacements(): ValidationResult {
    // Implementation will go here
  }
}
